[WIP] ALGOR-33: crypto support

+++ model/src/main/scala/com/github/ppotseluev/algorate/Order.scala
@@ -5,7 +5,7 @@ import enumeratum._
 
 case class Order(
     instrumentId: InstrumentId,
-    lots: Int,
+    lots: Double,
     operationType: OperationType,
     details: Order.Details,
     info: Order.Info
